clean unused

+++ core/src/main/java/com/example/core/utils/DataMapper.kt
@@ -45,7 +45,7 @@ object DataMapper {
 
     fun userSetFavoriteUser(data : UserDetail): FavoriteUser =
         FavoriteUser(
-            data.id.toInt(),
+            data.id,
             data.nickName.toString(),
             data.userName,
             data.avatarUrl,
